Move internal classes to subfolder, fix namespace
